styled & updated ProjectSelectContainer

+++ client-user/src/containers/ProjectSelectContainer.js
@@ -4,9 +4,13 @@ import styled, {css} from 'styled-components';
 import * as API from '../ServerApi'
 import InputDialog from "../components/dialogs/InputDialog";
 import ShareDialog from "../components/dialogs/ShareDialog";
-import {Button} from "../components/base/Button";
-import LinkedText from "../components/base/LinkedText";
+import {
+    Button, greenFilledTheme, greyBorderTheme, redBorderTheme, redFilledTheme,
+    redTheme
+} from "../components/base/Button";
+import {Link, greenHoverTheme} from "../components/base/Link";
 import ButtonIcon from "../components/base/ButtonIcon";
+import {MessageBoxDialog, MessageBoxButtonsEnum, MessageBoxResultEnum} from "../components/dialogs/MessageBoxDialog";
 
 
 const SubInfo = styled.div`
@@ -15,7 +19,9 @@ const SubInfo = styled.div`
 `;
 
 
-const ProjectListItem = styled.div`
+const ProjectListItem = styled(Link).attrs({
+    theme: greenHoverTheme
+})`
     padding-top: 10px;
     padding-bottom: 10px;
     border-bottom: 1px solid #ccc;
@@ -52,6 +58,10 @@ export default class ProjectSelectContainer extends React.Component {
         return {};
     }
 
+
+    projects = [];
+
+
     constructor(props){
         super(props);
 
@@ -60,11 +70,11 @@ export default class ProjectSelectContainer extends React.Component {
         };
     }
 
-    projects = [];
 
     componentDidMount() {
         API.getProjects(this.props.user.token ,this.handleProjectsReceived, this.handleError);
     }
+
 
     handleProjectsReceived = (body) => {
         this.projects = [];
@@ -84,6 +94,7 @@ export default class ProjectSelectContainer extends React.Component {
         });
     };
 
+
     handleCreateProjectClick = () => {
         this.props.showDialog(<InputDialog
             title="Create Project"
@@ -95,9 +106,11 @@ export default class ProjectSelectContainer extends React.Component {
         />);
     };
 
+
     handleCreateProject = (title) => {
         API.createProject(this.props.user.token, title, (body) => this.handleProjectCreated(title, body.projectId), this.handleError);
     };
+
 
     handleProjectCreated = (title, id) => {
         // add project to projects list
@@ -117,42 +130,75 @@ export default class ProjectSelectContainer extends React.Component {
         })
     };
 
-    handleProjectClick = (project) => {
+
+    handleProjectClick = (project) => (e) => {
         this.props.onProjectClick(project.id, project.title, project.permissions);
+        e.preventDefault();
     };
 
-    handleProjectDelete = (project) => {
-        // TODO add yes no dialog
-        API.deleteProject(this.props.user.token, project.id, () => this.handleProjectDeleted(project.id), this.handleError);
+
+    handleProjectDeleteClick = (project) => (e) => {
+        const handleResult = (result) => {
+            this.props.showDialog(null);
+
+            if (result === MessageBoxResultEnum.OK_YES)
+                API.deleteProject(this.props.user.token, project.id, () => this.handleProjectDeleted(project.id), this.handleError);
+        };
+
+        this.props.showDialog(
+            <MessageBoxDialog
+                title="Delete Project"
+                onResult={handleResult}
+                buttons={MessageBoxButtonsEnum.YES_NO}
+                buttonOkYesText="Delete"
+                buttonOkYesTheme={redFilledTheme}
+                buttonCancelNoText="Cancel"
+                buttonCancelNoTheme={greyBorderTheme}
+            >
+                Do you really want to delete "{project.title}"?
+            </MessageBoxDialog>
+        );
+
+        e.preventDefault();
+        e.stopPropagation();
     };
+
 
     handleProjectDeleted = () => {
         API.getProjects(this.props.user.token, this.handleProjectsReceived, this.handleError);
     };
 
-    handleProjectGetShares = (project) => {
+
+    handleProjectShareClick = (project) => (e) => {
         this.props.showDialog(<ShareDialog
             title="Share Project"
             projectId={project.id}
             onCancel={() => this.props.showDialog(null)}
             user={this.props.user}
         />);
+        e.preventDefault();
+        e.stopPropagation();
     };
+
 
     handleError = (msg) => {
         alert(msg);
     };
+
 
     getProjectView = () => {
         let list = [];
         // TODO make cool styling
         for (let p of this.state.projects) {
             list.push(
-                <ProjectListItem key={p.id}>
+                <ProjectListItem
+                    key={p.id}
+                    onClick={this.handleProjectClick(p)}
+                >
                     <ProjectListItemInner primary left>
                         <div>
                             <div>
-                                <LinkedText label={p.title} onClick={() => this.handleProjectClick(p)}/>
+                                {p.title}
                             </div>
                             <SubInfo>
                                 Shared by: {p.ownerName} ({p.ownerEmail})
@@ -161,14 +207,15 @@ export default class ProjectSelectContainer extends React.Component {
                     </ProjectListItemInner>
                     <ProjectListItemInner right>
                         <Button
-                            onClick={() => this.handleProjectGetShares(p)}
+                            onClick={this.handleProjectShareClick(p)}
                             marginLeft
                             marginRight
                         >
-                            Shares
+                            Share
                         </Button>
                         <Button
-                            onClick={() => this.handleProjectDelete(p)}
+                            onClick={this.handleProjectDeleteClick(p)}
+                            theme={redBorderTheme}
                         >
                             Delete
                         </Button>
@@ -180,10 +227,14 @@ export default class ProjectSelectContainer extends React.Component {
         return list;
     };
 
+
     render() {
         return (
             <div>
-                <Button onClick={this.handleCreateProjectClick}>
+                <Button
+                    onClick={this.handleCreateProjectClick}
+                    theme={greenFilledTheme}
+                >
                     Create Project
                 </Button>
                 {this.getProjectView()}
